requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
dateparser==1.1.7
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
joblib==1.2.0
multidict==6.0.4
numpy==1.24.2
psycopg2-binary==2.9.5
pycryptodome==3.17
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
websocket==0.2.1
websockets==10.4
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
